Removed mining phase + added stuff
There is no mining phase anymore.
Players will appear with a random ship on the mod and when diying they will be able to chose a ship for 10 seconds (10 seconds of invulnerability + idle mod) and then will be randomly teleported in the map. 
Ship balance has been worked, however some ships were Tier 7 and Tier 5 so are not balanced yet.